dev: reflect design change of chat identifiers to a json file, and minor chat packet initial tests

+++ src/main/java/com/aquamancer/warlordsbufficons/statuses/Status.java
@@ -62,7 +62,7 @@ public class Status {
     public long sync(Map.Entry<String, Integer> status) {
         return 0;
     }
-    public void setRemainingDuration(long duration) {
+    public void setRemainingDuration(int duration) {
         if (this.initialDuration < duration) {
             this.initialDuration = duration;
         }
